+++ addon/mixins/service-mixin.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
+import fetch from 'ember-network/fetch';
 
 export default Ember.Mixin.create({
-  arcgisAjax: Ember.inject.service(),
   session: Ember.inject.service('session'),
 
   hostAppConfig: Ember.computed(function () {
@@ -39,14 +39,40 @@ export default Ember.Mixin.create({
   }),
 
   encodeForm (form = {}) {
-    Ember.merge(form, this.get('defaultParams'));
+    // Ember.merge(form, this.get('defaultParams'));
     return Object.keys(form).map((key) => {
       return [key, form[key]].map(encodeURI).join('=');
     }).join('&');
   },
 
   /**
-   * Centralized Request with ArcGIS Payload checking
+   * Fetch does not reject on non-200 responses, so we need to check this manually
+   */
+  checkStatusAndParseJson (response) {
+    var error;
+    if (response.status >= 200 && response.status < 300) {
+      // check if this is one of those groovy 200-but-a-400 things
+      return response.json().then((json) => {
+        if (json.error) {
+          // cook an error
+          error = new Error(json.error.message);
+          error.response = response;
+          Ember.debug('Error in response:  ' + json.error.message);
+          throw error;
+        } else {
+          return json;
+        }
+      });
+    } else {
+      // Response has non 200 http code
+      error = new Error(response.statusText);
+      error.response = response;
+      throw error;
+    }
+  },
+
+  /**
+   * Fetch based request method
    */
   request (url, options) {
     let opts = options || {};
@@ -54,10 +80,22 @@ export default Ember.Mixin.create({
     if (opts.method && opts.method === 'POST') {
       // if we are POSTing, we need to manually set the content-type because AGO
       // actually does care about this header
-      opts.headers = {
-        'Accept': 'application/json',
-        'Content-Type': 'application/x-www-form-urlencoded'
-      };
+      if (!opts.headers) {
+        opts.headers = {
+          'Accept': 'application/json, application/xml, text/plain, text/html, *.*',
+          'Content-Type': 'application/x-www-form-urlencoded'
+        };
+      }
+      // if a body was passed, we need to set the content type to multipart
+      if (opts.body) {
+        delete opts.headers['Content-Type'];// = 'multipart/form-data';
+      }
+
+      // if we have a data, create a formData from it
+      if (opts.data) {
+        var form = this.encodeForm(opts.data);
+        opts.body = form;
+      }
     }
 
     // append in the token
@@ -71,6 +109,7 @@ export default Ember.Mixin.create({
       }
     }
 
-    return this.get('arcgisAjax').request(url, options);
+    return fetch(url, opts)
+      .then(this.checkStatusAndParseJson);
   }
 });