+++ tests/dummy/app/items/new/controller.js
@@ -24,6 +24,11 @@ export default Ember.Controller.extend({
       this.get('itemsService').create(item)
         .then((resp) => {
           this.transitionToRoute('items.index');
+        }, (error) => {
+          alert('Error Creating Item: ' + error.message);
+        })
+        .catch((error) => {
+          alert('Error Creating Item 2: ' + error.message);
         });
     },
     cancel: function () {